Initial Commit: Custom MOVG Library for Moving Average with Outlier Elimination
Added the initial version of the MOVG (Moving Average with Outlier Elimination) library. This library is designed to compute a moving average of a specified window size while eliminating outliers beyond a given threshold. The class includes methods for initializing the MOVG instance and computing the moving average with outlier elimination. The library is intended to be used for smoothing sensor data and reducing the impact of outliers.