Move MOVG.h to Arduino IDE library location
This commit relocates the `MOVG.h` file to the appropriate library location within the Arduino IDE on my local machine. By moving the file to the library directory, it ensures that the MOVG class can be easily accessed and included in Arduino sketches. This organization follows best practices for managing Arduino libraries, enhancing code modularity, and simplifying project dependencies.