Modify file naming based on SD card file count
In this commit, the file naming logic in the Arduino sketch has been modified to generate filenames based on the count of existing files in the SD card. The setup function now counts the number of files in the SD card's root directory and uses that count to construct a new filename for the data file. The filename format is "00000000.csv", where "00000000" represents the count of existing files. This ensures unique and sequential filenames for each data file constructed on the SD card.