Add overloaded function computeMOVG without parameters
This commit introduces an overloaded version of the `computeMOVG` function in the `MOVG` class. The new function does not require any parameters and calculates the Moving Average (MOVG) using the existing window size and outlier threshold. This addition enhances the usability of the class by providing a convenient way to compute the moving average without specifying a new value.