Add method to add values without computing moving average
This commit introduces a new method `addValue` to the `MOVG` class, allowing values to be added to the internal array without immediately computing the moving average. This enhancement provides more flexibility in managing the data before computing the moving average, improving the usability of the class. Additionally, the `computeMOVG` function has been modified to call `addValue` internally before computing the moving average.